requirements.txt
fastapi
uvicorn[standard]
google-generativeai
faiss-cpu
python-dotenv
tiktoken
pydantic
typing-extensions